+++ src/solver.ts
@@ -61,8 +61,9 @@ export default class Solver {
     let best = { loss: Infinity };
     let counter = 0;
     while (best.loss > this.options.acceptanceLossPercentage) {
-      const initial: FilterValuesArray = [50, 20, 3750, 50, 100, 100];
-      const result: SPSAPayload = this.spsa(A, a, c, initial, 1000);
+      const initialFilterValues: FilterValuesArray = [50, 20, 3750, 50, 100, 100];
+      const result: SPSAPayload = this.spsa({ A, a, c, values: initialFilterValues, maxTriesInLoop: 1000 });
+
       if (result.loss < best.loss) {
         best = result;
       }
@@ -81,20 +82,72 @@ export default class Solver {
     const c = 2;
     const A1 = A + 1;
     const a = [0.25 * A1, 0.25 * A1, A1, 0.25 * A1, 0.2 * A1, 0.2 * A1];
-    return this.spsa(A, a, c, wide.values, wide.called);
+    return this.spsa({
+      A,
+      a,
+      c,
+      values: wide.values,
+      maxTriesInLoop: 500,
+      called: wide.called,
+    });
   }
 
-  private spsa(
-    A: number,
-    a: number[],
-    c: number,
-    // TODO: check this any
-    values: FilterValuesArray,
+  /**
+   * Returns final value based on the current filter order
+   * to get the order, please check the returned value
+   * in `css()` method
+   *
+   * @private
+   * @param {number} value
+   * @param {number} idx
+   * @returns {number}
+   * @memberof Solver
+   */
+  private fixValueByFilterIDX(value: number, idx: number): number {
+    let max = 100;
+
+    // Fixing max, minimum and value by filter
+    if (idx === 2 /* saturate */) {
+      max = 7500;
+    } else if (idx === 4 /* brightness */ || idx === 5 /* contrast */) {
+      max = 200;
+    }
+
+    if (idx === 3 /* hue-rotate */) {
+      if (value > max) {
+        value %= max;
+      } else if (value < 0) {
+        value = max + (value % max);
+      }
+    }
+    // Checking if value is below the minimum or above
+    // the maximum allowed by filter
+    else if (value < 0) {
+      value = 0;
+    } else if (value > max) {
+      value = max;
+    }
+    return value;
+  }
+
+  private spsa({
+    A,
+    a,
+    c,
+    values,
+    maxTriesInLoop = 500,
     called = 0,
-  ): SPSAPayload {
+  }: {
+    A: number;
+    a: number[];
+    c: number;
+    // TODO: check this any
+    values: FilterValuesArray;
+    maxTriesInLoop: number;
+    called?: number;
+  }): SPSAPayload {
     const alpha = 1;
     const gamma = 0.16666666666666666;
-    const maxTriesInLoop = 500;
 
     let best = null;
     let bestLoss = Infinity;
@@ -103,34 +156,11 @@ export default class Solver {
     const highArgs = new Array(6) as FilterValuesArray;
     const lowArgs = new Array(6) as FilterValuesArray;
 
-    function fix(value: number, idx: number): number {
-      let max = 100;
-      // Fixing max, minimum and value by filter
-      if (idx === 2 /* saturate */) {
-        max = 7500;
-      } else if (idx === 4 /* brightness */ || idx === 5 /* contrast */) {
-        max = 200;
-      } else if (idx === 3 /* hue-rotate */ && value > max) {
-        value %= max;
-      } else if (idx === 3 /* hue-rotate */ && value < 0) {
-        value = max + (value % max);
-      }
-      // Checking if value is below the minimum or above
-      // the maximum allowed by filter
-      else if (value < 0) {
-        value = 0;
-      } else if (value > max) {
-        value = max;
-      }
-
-      return value;
-    }
-
     // Size of all CSS filters to be applied to get the correct color
     const filtersToBeAppliedSize = 6;
 
-    for (let k = 0; k < maxTriesInLoop; k++) {
-      const ck = c / Math.pow(k + 1, gamma);
+    for (let key = 0; key < maxTriesInLoop; key++) {
+      const ck = c / Math.pow(key + 1, gamma);
       for (let i = 0; i < filtersToBeAppliedSize; i++) {
         deltas[i] = Math.random() > 0.5 ? 1 : -1;
         highArgs[i] = values[i] + ck * deltas[i];
@@ -140,8 +170,8 @@ export default class Solver {
       const lossDiff = this.loss(highArgs) - this.loss(lowArgs);
       for (let i = 0; i < filtersToBeAppliedSize; i++) {
         const g = (lossDiff / (2 * ck)) * deltas[i];
-        const ak = a[i] / Math.pow(A + k + 1, alpha);
-        values[i] = fix(values[i] - ak * g, i);
+        const ak = a[i] / Math.pow(A + key + 1, alpha);
+        values[i] = this.fixValueByFilterIDX(values[i] - ak * g, i);
       }
 
       const loss = this.loss(values);